refactor(server): Utilize "usize" for stream ID and topic ID
According to #2198, align ID types of stream and topic with the target pointer
size "usize" to avoid churny casts and platform-specific inconsistencies.

Some parts outside the scope of server needs to be changed due to the depency
for structs under iggy common.

+++ core/ai/mcp/src/service/requests.rs
@@ -411,7 +411,7 @@ pub struct Permissions {
     pub global: Option<GlobalPermissions>,
 
     #[schemars(description = "stream permissions (optional)")]
-    pub streams: Option<HashMap<u32, StreamPermissions>>,
+    pub streams: Option<HashMap<usize, StreamPermissions>>,
 }
 
 #[derive(Debug, Deserialize, JsonSchema)]
@@ -468,7 +468,7 @@ pub struct StreamPermissions {
     pub send_messages: Option<bool>,
 
     #[schemars(description = "topics permissions (optional)")]
-    pub topics: Option<HashMap<u32, TopicPermissions>>,
+    pub topics: Option<HashMap<usize, TopicPermissions>>,
 }
 
 #[derive(Debug, Deserialize, JsonSchema)]
